Add files via upload
Added support to check that the User's IVs are above the minimum IVs, and added a check to ensure we won't output NaN (for example if the minLvl pushes the CP above the league limit.  And will check if the User's Inputs are impossible for a given mon and adjust them up to the minimum allowed levels (1 IVs /  1 IV floor for legendaries via special trades, and 10 IVs / 10 IV floor for non-tradeable quest/raid mon such as Mew, Celebi, Jirachi, and Deoxys).

Also added code to modify the webpage title to reflect the user's search parameters for SEO/usage metric tracking

Also fixed an error where A-Grimer wasn't being output for A-Muk (space present in availableMon).